Using a counter instead of accumulator to start_time.
There was a but where an incorrect number of samples would be generated because of the increasing rounding error when incrementing start_time.

For instance, by running bitrate with -m1 you get N samples, running with -m10 should be ~10*N (with exception for the last 0-N samples). The scaling should always be ~10 if continuing to scale -m by 10. With the rounding error in place it generated either to many but mostly to few samples at certain sampling frequencies (like 10kHz).